+++ CHANGELOG.rst
@@ -2,7 +2,11 @@ Changelog
 =========
 
 
-- Add ``Select`` schema helper.
+- Add schema helpers:
+
+  - ``Select``
+  - ``Use``
+
 - Include execption class name in error message returned by ``As``.
 - Always return a ``dict`` when parsing from dictionary schemas instead of trying to use the source data's type as an initializer. (**breaking change**)